+++ src/lxspi/miso.rs
@@ -1,5 +1,5 @@
 #[doc = "Reader of register MISO"]
-pub type R = crate::R<u32, super::MISO>;
+pub type R = crate::R<u8, super::MISO>;
 #[doc = "Reader of field `miso`"]
 pub type MISO_R = crate::R<bool, bool>;
 impl R {
