+++ src/messible/out.rs
@@ -1,5 +1,5 @@
 #[doc = "Reader of register OUT"]
-pub type R = crate::R<u32, super::OUT>;
+pub type R = crate::R<u8, super::OUT>;
 #[doc = "Reader of field `out`"]
 pub type OUT_R = crate::R<u8, u8>;
 impl R {
