+++ src/timer0/reload1.rs
@@ -1,10 +1,10 @@
 #[doc = "Reader of register RELOAD1"]
-pub type R = crate::R<u32, super::RELOAD1>;
+pub type R = crate::R<u8, super::RELOAD1>;
 #[doc = "Writer for register RELOAD1"]
-pub type W = crate::W<u32, super::RELOAD1>;
+pub type W = crate::W<u8, super::RELOAD1>;
 #[doc = "Register RELOAD1 `reset()`'s with value 0"]
 impl crate::ResetValue for super::RELOAD1 {
-    type Type = u32;
+    type Type = u8;
     #[inline(always)]
     fn reset_value() -> Self::Type {
         0
@@ -20,7 +20,7 @@ impl<'a> RELOAD_W<'a> {
     #[doc = r"Writes raw bits to the field"]
     #[inline(always)]
     pub unsafe fn bits(self, value: u8) -> &'a mut W {
-        self.w.bits = (self.w.bits & !0xff) | ((value as u32) & 0xff);
+        self.w.bits = (self.w.bits & !0xff) | ((value as u8) & 0xff);
         self.w
     }
 }
