+++ src/touch/i.rs
@@ -1,5 +1,5 @@
 #[doc = "Reader of register I"]
-pub type R = crate::R<u32, super::I>;
+pub type R = crate::R<u8, super::I>;
 #[doc = "Reader of field `i`"]
 pub type I_R = crate::R<u8, u8>;
 impl R {
