+++ src/usb/in_data.rs
@@ -1,10 +1,10 @@
 #[doc = "Reader of register IN_DATA"]
-pub type R = crate::R<u32, super::IN_DATA>;
+pub type R = crate::R<u8, super::IN_DATA>;
 #[doc = "Writer for register IN_DATA"]
-pub type W = crate::W<u32, super::IN_DATA>;
+pub type W = crate::W<u8, super::IN_DATA>;
 #[doc = "Register IN_DATA `reset()`'s with value 0"]
 impl crate::ResetValue for super::IN_DATA {
-    type Type = u32;
+    type Type = u8;
     #[inline(always)]
     fn reset_value() -> Self::Type {
         0
@@ -20,7 +20,7 @@ impl<'a> DATA_W<'a> {
     #[doc = r"Writes raw bits to the field"]
     #[inline(always)]
     pub unsafe fn bits(self, value: u8) -> &'a mut W {
-        self.w.bits = (self.w.bits & !0xff) | ((value as u32) & 0xff);
+        self.w.bits = (self.w.bits & !0xff) | ((value as u8) & 0xff);
         self.w
     }
 }
