+++ src/usb/setup_ctrl.rs
@@ -1,10 +1,10 @@
 #[doc = "Reader of register SETUP_CTRL"]
-pub type R = crate::R<u32, super::SETUP_CTRL>;
+pub type R = crate::R<u8, super::SETUP_CTRL>;
 #[doc = "Writer for register SETUP_CTRL"]
-pub type W = crate::W<u32, super::SETUP_CTRL>;
+pub type W = crate::W<u8, super::SETUP_CTRL>;
 #[doc = "Register SETUP_CTRL `reset()`'s with value 0"]
 impl crate::ResetValue for super::SETUP_CTRL {
-    type Type = u32;
+    type Type = u8;
     #[inline(always)]
     fn reset_value() -> Self::Type {
         0
@@ -30,7 +30,7 @@ impl<'a> RESET_W<'a> {
     #[doc = r"Writes raw bits to the field"]
     #[inline(always)]
     pub fn bit(self, value: bool) -> &'a mut W {
-        self.w.bits = (self.w.bits & !(0x01 << 5)) | (((value as u32) & 0x01) << 5);
+        self.w.bits = (self.w.bits & !(0x01 << 5)) | (((value as u8) & 0x01) << 5);
         self.w
     }
 }
